Bing Search/requirements.txt
adal==1.2.5
azure-cognitiveservices-search-entitysearch==2.0.0
azure-cognitiveservices-search-websearch==2.0.0
azure-common==1.1.25
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2
idna==2.10
isodate==0.6.0
msrest==0.6.19
msrestazure==0.6.4
oauthlib==3.1.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.11
